Add vimbundle
The local vimbundle contains all of the Plug plugins to use with
vim that are custom from dotmatrix.

I added ctrlp because I like it. And I disabled all dotmatrix plugins to
increase the speed of vim. I'll re-enable the necessary ones as I pair
with folks and find out which ones we actually use.