Tweak tmux split & window bindings paths
This makes it so that the bigcartel/dotmatrix new window in the current
pane's path is removed to restore the default behavior. It also makes it
so the default tmux vertical and horizontal pane splits use the current
path.

This reflects the way I prefer to use tmux, with each window being a

different project and each pane being a split for that project.